Move LinearMixedModel to stats module (#4083)
* wip

removed max_rank

lots of stuff

wip

wip

* svd update

* added kinship example, improvements, checks

* small fixes

* move LinearMixedModels to stats, doc fixes
