+++ python/hail/docs/methods/stats.rst
@@ -10,12 +10,10 @@ Statistics
 
     row_correlation
     linear_regression
-    LinearMixedModel
     logistic_regression
     pca
 
 .. autofunction:: row_correlation
 .. autofunction:: linear_regression
-.. autoclass:: LinearMixedModel
 .. autofunction:: logistic_regression
 .. autofunction:: pca
