+++ python/hail/methods/__init__.py
@@ -9,7 +9,6 @@ from .statgen import linear_regression, logistic_regression, skat, identity_by_d
     split_multi_hts, balding_nichols_model, ld_prune, row_correlation, ld_matrix
 from .qc import sample_qc, variant_qc, vep, concordance, nirvana, summarize_variants
 from .misc import rename_duplicates, maximal_independent_set, filter_intervals, window_by_locus
-from .linear_mixed_model import LinearMixedModel
 
 __all__ = ['trio_matrix',
            'linear_regression',
@@ -58,7 +57,6 @@ __all__ = ['trio_matrix',
            'de_novo',
            'filter_alleles',
            'filter_alleles_hts',
-           'LinearMixedModel',
            'summarize_variants',
            'row_correlation',
            'ld_matrix'
